[FLOWCONTROL] Control the flow from the jobs to the Workerpool. (#1717)
* [FLOWCONTROL] Control the flow from the jobs to the Workerpool.

It is observed that some plugins require a lot of chattiness to realize
functionality. A ood example of this is the PersitentStore plugin from the
RDKServices. Also the call duration of each issued JSONRPC call is lengthy.
This cuases an flood of messages on the Workerpool. The lengthyness of the
call is due to the slow backend database where the key values are stored.
A lightweight alternative is the Dictonary available in the
ThunderNanoServices.

This mechanism introduced here is to throttle down plugins that consume too
workerpool reources from the framework. To prevent chatty and bloated plugins,
like the Persistenttore plugin, to flood the system. With this mechanism in
place you can limit the number of JSONRPC requests placed in the workerpool,
and issued on a channel.

* Update PluginServer.h

Remove the parts already in the ControlData, it moved to a dedicated class ;-)