Development/secure socket port (#1795)
* [ASSERT] If Thunder is to avoid a crash, at least report it using an ASSERT!

* [SECURESOCKET] Start enabling the possibility to run a SecureSocketPort server.

Sofar Thunder supported SecureSocket clients where the intiative to enable secure connections
was only based on creating a socket as a client. Tests are being written that require a secure
socket server. This requires a sifferent set of calls and revealed an issue in the SocketPort.

The server accepts the incoming socket. The bug was that the incoming socket, as it is already
open and connected would not call the "Initialize()" method before it was added to the Resource
Monitor. This has been fixed in SocketPort.cpp.

Than this new incoming connection *must* follow a different path in the interaction with the
opensll library. Already added new states like, ACCEPTING and CONNECTING, in stead of EXCHANGE
however, testing it with a server build using the Thunder fnctionality still fails. It is
probably due to the functionality/initialization that has to take place in the Initialize()
method of the SecureScketPort where the SSL context is being created. The diffrentiation
between a client socket and a server socket is already added to the Initialize, I guess it
has to be tweaked.

The client functionality of the SecureSocketPort has been tested with a server on the internet
and is still working oke, so Client functionality has *not* been broken with these changes!