[SINGLETON] Add a UniqueType, whihc is a ref counted singleton. (#1849)
* [SINGLETON] Add a UniqueType, whihc is a ref counted singleton.

If the lifetme of singleton should be limited by its lifetime, there is
a new template in town that will generate a singleton *only* if there is
a user for it. If there is no user, it will be destructed and recreated
if it is needed later in time.

Maybe we could call this a ref counted singleton! Use case is the Backend
to render surfaces (in the compositor). As soon as there is a client, each
client requires a Backend. The backend must be a singleton over all existing
clients but once there is no client anymore, the backend should be destructed
as well. The template in this PR relaizes that functionality.

* [WINDOWS] Fixes for 64Bits platforms.