[Dangling] Improve the offloading of Dangling proxies notification. (#1918)
* [Dangling] Improve the offloading of Dangling proxies notification.

* move the proxies while notifying dangling

* Remove DanglingNotifier

* Revert back to vector

* [PROPOSAL] Addressing the locking of the observers and the single add..

Most of the notifications currently in the Thunder system are safe to
destruct on the unregister. This means that it is guaranteed that no
notification is received anymore after the unregsister is completed.
This makes coding easy as it means that the noification sinks can be
composits of a class. Once the unrregister in the destructror is
called, the sink can be destroyed to.

If the unregister can *not* guarantee that the notification is nolonger
called after the unregister took place, destructions of a sink becomes
more cumbersome.

The curret setup (copying the observer list) breakes the rule that
after the unregister is called, no more notification can be received.
The copied (and hopefully addreffed interface) could potentially be
called after the Unregister took the notification interface uit of
the original list.

This means that the construction of the Sink must adhere to other
rules (it can not be a composit of another class, a typical usecase).

As this will probably lead to sporadic, hard to reproduce, crashes,
it is prefered to stick to the rule:
"After an Unregister returns, no more notifications will be issued"
as this makes life and coding of the consumer of notifications
simpler.
Unfortunatly it makes the life and coding of the producer of the
notification harder. In the proposal you find here, I hope to have
solved it by using two locks. There is now a dedicated lock for the
dangling proxies. This way the producer (potentially the resource
monitor thread) is now locking the access to the shared list of
dangling proxies and nolonger the same lock as we use to protect
the observer list for the notifications. This lock for the
notifications (_adminLock) is now not used in the method that
adds the deadproxies to the shared pool and thus i can currently
only be blocked if the new shared resource lock is takes. That new
lock however, is only briefly taken, only during the deterministic
removal of 1 element out of the shared resources (dangling proxies).
No other actions that could potentially require a communication
thread is within this scope, reslving the deadlock from the analyses.

Next to this, the seup as it was, was adding one dangling proxy at
a time and firing the job on each insertion. I think that was a bit
of overkill if we for example had to add ten dangling proxies. We
know up front how many dangling proxie we have so lets add the all
so that the job starting is done only one per deleted channel.

* Make the Dangling method non-pure virtual to maintain backwards compatibility

* Remove the std::move to fix compilation error

* Remove the std::move to fix compilation error

* Fix warning

---------

Co-authored-by: Pierre Wielders <pierre@wielders.net>