[Backport] Offloading of Dangling proxies notification (#1923)
* [Backport] Offloading of Dangling proxies notification
When Communication thread receives a broken pipe,
dangling notification is sent to the CommunicationServer.
This will make the comm thread wait on the notifier lock when
activation/deactivation of any plugin is in progress and
the notifier lock for IPlugin::INotification was already taken
The notification to ICOMLink::INotification subscribers is now
offloaded to a Job in CommunicationServer.

* Fix the missing Release in Dangling Notification