[COMRPC] Close the socket when a COMRPC call timesout (#1927)
* When a COMRPC call times out, on a channel, most of the times the caller
gives up and continues without handling the error. But the other side may
still send a response when its not expected. This causes the response message to be
processed for a different outbound method from Thunder on the wrong time
resulting in a memory corruption and crash. This change closes the SocketPort
when there is a COMRPC timeout thereby avoiding subsequent outbound messages on the
channel and also triggering fault tolerance code on the communication
thread to kick-in and gracefully shudown the plugin.

* fix logging

---------

Co-authored-by: Pierre Wielders <pierre@wielders.net>