[SERVICE DISCOVERY] Change the way the services are discovered from c… (#1939)
* [SERVICE DISCOVERY] Change the way the services are discovered from centralized to decentralized.

As the service discovery was centralized, the build-up of the centralized list had to be protected. Service registration
in the centralized setup was done at the moment, an so (dll) is loaded.

The dyamic loader of linux (dl) has a mutex to protect other DL calls to happen while an open or close is in progress. This
means that if a library is opend/loaded, the dl mutex (Lock A) is taken. During the loading the the statics in the module are run,
as part of this run, the services are registered in the centralized services store, which in turn use a mutex to protect
the list of available service (Lock B)

Now if a service is constructed, the centralized services are consulted, to make sure they are not removed, while we creat a service,
the list is protected with a mutex (Lock B), if during the instatiation of this service, other libraries need to be loaded, the
linux dynamic loader will call dlopen which in turn lock any other action for dl by taking a mutex (Lock A).

If you have a serious QA department that tests properly (automated test that continously start and stop Thunder) you will see that
the slim change of the multithreaded Thunder startup might cause a coincident "Load Library" and "Create Service" between the taking
of the first lock and be eempted before the second lock was taken, that will lead to a classical ABBA lock.

To resolve it, initially we used a paradigm to make sure that lock B was *always* taken before lock A was taken. Unfortunatly this
brings a bit of implementation knowledge to the developer of a plgin that is very hard to explain nor to check / validate. So this
might lead to unexpected deadlocks (with a very slim chance) in the future if the B lock was not always taken before the A lock.
The goal of this PR is to remove the use of the, till now, nessecary lock B (Lock A is in the linux dynamic loader and can not be
controlled by the Thunder coe). This is done by using a decentralized approach. The services (and there factory implementations)
are nolonge "registered" centrally, but in stead the Library (if loaded) is now searched for a "IService" entry point. This Iservice
is now enhancedwith a Next(), which allows to iterate over all services in this library. This chain is build during the loading of the
library (Single thread and protected by the dl mutex ;-) ), Once the library is loaded, the dl's lifetime is protected wth the
dl's handle refernce count (implemented in Core::Library) as the lifetime for creation is now protected by the ref count on the library
and the list is not subject to change anymore, we are nolonger in the need of "Lock B" which resolves the "under the hood" knowledge!

* Update Library.h

* Update Library.h

* Update Library.h

* Update Services.h

* Update Services.cpp

* Update Library.h

* Update PluginServer.h

* Update Services.cpp

* Update Services.h

* Update Services.h

* Update Library.h

* Update Library.h

* Update Library.h

* [REVIEW] Adressing review remarks.

* [TESTING] On linux the load of an address for the application did not work :-)

* Update Services.h

* Update Services.cpp

* Update Communicator.h

* [LOADLIBRARY] Refactored the LoadLibrary functionality.

* Update Services.h

* Update PluginServer.h

---------

Co-authored-by: Mateusz Daniluk <121170681+VeithMetro@users.noreply.github.com>