[METADATADISCOVERY] Make the Metadata discover configurable.
For automatic configuration of the subsystems and plugin version retrieval, thunder loads all
plugins into memory at startup without starting them. This allows for a proper build up of the
subsystem dependencies (read from the build in code of the plugin) and allows for retrieval of
plugin versions, without the plugins being activated.
This feature however, moves the additional loading time of each plugin, into the Thunder startup
timer. If squashfilesystem (assumption this is the rootcasue) is used and you have a great deal
of plugins that are not activated at startup, increase the Thunder Plugin time by 6 Seconds.
Measurements using an ext4 filesystem show (using 150 reference stack plugins) that the additional
loading time is increased by 600mS.

To reduce the startup time of the software it was requested by Comcast to make this feature
configurable. With this PR, there is a flag in the config called "discovery". By default it is true
Comcast can, in their build, configure this to false. If configured to false, the plugins are
nolonger preloaded to extracte the metadata from the plugins, effectively reducing the initial
startup time of Thunder.

Note: If disabled, the subsystems functionality, is nolonger taken from the plugin and if needed
      must be placed in the plugin config, manual labour, error prone but still available.
      Versions of plugins will not give any output as long as the plugin is not loaded, at least
      once. So the plugin must be at least started once to get this version info.
      The startup time is still required at first startup, so the feature only "moves" the startup
      time from Thunder to whenever the plugin is activated for the first time!

      Bottonlime: From Thunder perspective, it is *not* recommended to turn of the "discovery"!