[RACECONDITION] Racecondition on destruction of a socket if the Closed() was executed.
If the Closed() method, was run, it would set the m_State = 0, which meant that the
SocketPort class was classed and eligable for destruction. In a race condition, this
signalling allowed a server to "cleanup" the children (Clients). In the unlucky case
that after the setting of the:

[ReourceMonitorThread]
SocketPort::Closed() {
    ...
    m_State -> 0
    ...
}
                                             [Server Cleanup Thread]
                                             Delete all Clients that are Closed()
SocketPort::Events() {
    ...
    m_State &= ~SocketPort::MONITOR;  CRASH!
    ...
}

This PR prevents this race condition from happening.

As this is in the core of Thunder, the biggest risk is the Lock in the
destructor. It might casue (if the socket is used incorrectly) to ABBA
locks. Request thorough testing !