If the Closed() method, was run, it would set the m_State = 0, which meant that the SocketPort class was closed and eligable for destruction. iThe CLosed() is always running on the ResourceMonitor thread. In a race condition, this setting of the m_State = 0 allowed a server to "cleanup"  the children (Clients). In the unlucky case that after the setting of the m_State =0, the kernel would preempt the ResourceMonitor thread for later continuation, the Server cleanup thread migh have killed the whole SocketPort object. Effectively any operation following the closed() on the ResourceMonitor threas that was preempted and continues would happen on a dead object!
[ReourceMonitorThread]
SocketPort::Closed() {
...
m_State -> 0
[preempted]

[Server Cleanup Thread]
Delete all Clients that are Closed()
[preemted]

[ResourceMonitorThread]
m_State &= ~SocketPort::MONITOR; CRASH!
...
}

This PR prevents this race condition from happening.

As this is in the core of Thunder, the biggest risk is the Lock in the destructor. It might casue (if the
socket is used incorrectly) to ABBA locks. Request thorough testing !
Also the ASSERT to validate if the CLosed() is indeed only run on the resource monitor thread is a risk!